pyproject.toml
[tool.poetry]
name = "api-introspection"
version = "0.1.0"
description = "Upload an OpenAPI spec, get natural-language summaries of endpoints, test them with mock data, and document issues."
authors = ["Maya Reinart <mayareinart@gmail.com>"]
license = "GNU GPLv3"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.12"}
uvicorn = {extras = ["standard"], version = "^0.34.2"}
openai = "^1.82.0"
pydantic = "^2.11.5"
loguru = "^0.7.3"
httpx = "^0.28.1"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
coverage = "^7.8.2"
ruff = "^0.11.11"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
python-dotenv = "^1.1.0"
python-multipart = "^0.0.20"
celery = "^5.5.2"
redis = "^6.1.0"
prance = "^25.4.8"
pyyaml = "^6.0.1"
openapi-spec-validator = "^0.7.1"


[tool.poetry.group.dev.dependencies]
celery-types = "^0.23.0"
types-pyyaml = "^6.0.12.20250516"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Same as Black.
line-length = 88
# Assume Python 3.10
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes by default.
select = ["E", "F", "I", "B", "C", "N", "UP", "ANN", "S", "A", "DTZ", "T20", "RET", "SLF", "SIM", "TID", "ARG", "PTH", "ERA", "PD", "PL", "TRY", "RUF"]
ignore = ["ANN101"]  # Ignore missing type annotations for self in methods

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",  # Allow assert in tests
    "PLR0913",  # Allow many arguments in test functions
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_default_fixture_loop_scope = "function"  # Set default event loop scope to function level
asyncio_mode = "auto"
markers = [
    "completion_e2e: marks tests that make real API calls to OpenAI (deselect with '-m \"not completion_e2e\"')"
]
addopts = "-m 'not completion_e2e'"
